+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Header to TypeScript
 - Footer to TypeScript
 - Form (checkbox, password, select, textarea) to TypeScript
+- Layout (container, layout, pagelayout, sidebar) to TypeScript
 - Widgets (posts, products, tags, about, instagram, newsletter, promobox, search, social, title) to TypeScript
 
 #### Features
@@ -24,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - posts getPosts limit and category defaults of null
 - posts getTags limit default of null
 - products getProducts limit and category defaults of null
+- Layout components (container, layout, pagelayout, sidebar) to TypeScript
 
 ## [0.0.0] - 2025-11-21